ltt_dashboard/requirements.txt
django==4.0.6
djangorestframework==3.13.1
django-extensions==3.1.5
django-uuid-upload-path==1.0.0
psycopg2==2.9.3
python-dotenv==0.20.0
django-environ==0.9.0
django-versatileimagefield==2.2
djangorestframework-simplejwt==5.2.0